+++ packages/flutter_query/lib/src/core/query_client.dart
@@ -3,7 +3,7 @@ import 'package:meta/meta.dart';
 import 'default_mutation_options.dart';
 import 'default_query_options.dart';
 import 'infinite_query_function_context.dart';
-import 'infinite_query_observer_options.dart';
+import 'infinite_query_options.dart';
 import 'mutation_cache.dart';
 import 'mutation_state.dart';
 import 'query.dart';