+++ packages/flutter_query/lib/src/core/query_observer.dart
@@ -6,7 +6,7 @@ import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 
 import 'infinite_query_function_context.dart';
-import 'infinite_query_observer_options.dart';
+import 'infinite_query_options.dart';
 import 'infinite_query_result.dart';
 import 'observable.dart';
 import 'query.dart';